Refactor html cleaner out into a module
