+++ src/ggrc/models/__init__.py
@@ -4,10 +4,13 @@
 # Maintained By: dan@reciprocitylabs.com
 
 import inspect
+import sqlalchemy as sa
 
-from ggrc.models.all_models import *  # noqa
-from ggrc import settings
 from ggrc import db
+from ggrc import settings
+from ggrc.models.reflection import SanitizeHtmlInfo
+from ggrc.models.all_models import *  # noqa
+from ggrc.utils import html_cleaner
 
 """All gGRC model objects and associated utilities."""
 
@@ -143,51 +146,11 @@ def init_session_monitor_cache():
 
 def init_sanitization_hooks():
   # Register event listener on all String and Text attributes to sanitize them.
-  import bleach
-  from HTMLParser import HTMLParser
-  import sqlalchemy as sa
-  from ggrc.models.reflection import SanitizeHtmlInfo
-  from .all_models import all_models
-
-  # Set up custom tags/attributes for bleach
-  bleach_tags = [
-      'caption', 'strong', 'em', 'b', 'i', 'p', 'code', 'pre', 'tt', 'samp',
-      'kbd', 'var', 'sub', 'sup', 'dfn', 'cite', 'big', 'small', 'address',
-      'hr', 'br', 'div', 'span', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'ul',
-      'ol', 'li', 'dl', 'dt', 'dd', 'abbr', 'acronym', 'a', 'img',
-      'blockquote', 'del', 'ins', 'table', 'tbody', 'tr', 'td', 'th',
-  ] + bleach.ALLOWED_TAGS
-  bleach_attrs = {}
-  attrs = [
-      'href', 'src', 'width', 'height', 'alt', 'cite', 'datetime',
-      'title', 'class', 'name', 'xml:lang', 'abbr'
-  ]
-
-  for tag in bleach_tags:
-    bleach_attrs[tag] = attrs
-
-  def cleaner(target, value, oldvalue, initiator):
-    # Some cases like Request don't use the title value
-    #  and it's nullable, so check for that
-    if value is None:
-      return value
-
-    parser = HTMLParser()
-    value = unicode(value)
-    while True:
-      lastvalue = value
-      value = parser.unescape(
-          bleach.clean(value, bleach_tags, bleach_attrs, strip=True)
-      )
-      if value == lastvalue:
-        break
-    return value
-
-  for model in all_models:
+  for model in all_models.all_models:
     attr_info = SanitizeHtmlInfo(model)
     for attr_name in attr_info._sanitize_html:
       attr = getattr(model, attr_name)
-      sa.event.listen(attr, 'set', cleaner, retval=True)
+      sa.event.listen(attr, 'set', html_cleaner.cleaner, retval=True)
 
 
 def init_app(app):
